Remove `hints` surface from `MainLayout`. Add `keyshortcuts` surface to FRE pages. Polish PluginHelp components for a11y & layout.
