+++ packages/apps/plugins/plugin-help/src/components/Shortcuts/ShortcutsSection.tsx
@@ -1,0 +1,16 @@
+//
+// Copyright 2024 DXOS.org
+//
+import React from 'react';
+
+import { descriptionText } from '@dxos/react-ui-theme';
+
+import { ShortcutsList } from './ShortcutsList';
+
+export const ShortcutsSection = () => {
+  return (
+    <section className={descriptionText}>
+      <ShortcutsList />
+    </section>
+  );
+};
