+++ packages/apps/plugins/plugin-help/src/components/Shortcuts/index.ts
@@ -4,3 +4,6 @@
 
 export * from './ShortcutsDialog';
 export * from './ShortcutsHints';
+export * from './ShortcutsList';
+export * from './ShortcutsSection';
+export * from './Key';
