+++ packages/apps/plugins/plugin-layout/src/components/ContentEmpty.tsx
@@ -4,6 +4,7 @@
 
 import React from 'react';
 
+import { Surface } from '@dxos/app-framework';
 import { useTranslation } from '@dxos/react-ui';
 import { descriptionText, mx } from '@dxos/react-ui-theme';
 
@@ -17,15 +18,18 @@ export const ContentEmpty = () => {
       className='min-bs-screen is-full flex items-center justify-center p-8'
       data-testid='layoutPlugin.firstRunMessage'
     >
-      <p
-        role='alert'
-        className={mx(
-          descriptionText,
-          'border border-dashed border-neutral-400/50 rounded-lg flex items-center justify-center p-8 font-normal text-lg',
-        )}
-      >
-        {t('first run message')}
-      </p>
+      <div role='none' className='grid place-items-center grid-rows-[min-content_min-content]'>
+        <p
+          role='alert'
+          className={mx(
+            descriptionText,
+            'place-self-stretch border border-dashed border-neutral-400/50 rounded-lg text-center p-8 font-normal text-lg',
+          )}
+        >
+          {t('first run message')}
+        </p>
+        <Surface role='keyshortcuts' />
+      </div>
     </div>
   );
 };
