+++ packages/apps/plugins/plugin-layout/src/components/MainLayout.tsx
@@ -117,14 +117,6 @@ export const MainLayout = ({ fullscreen, showHintsFooter, showComplementarySideb
           <Surface role='status' limit={1} />
         </div>
 
-        {/* Help hints. */}
-        {/* TODO(burdon): Make surface roles/names fully-qualified. */}
-        {showHintsFooter && (
-          <div className='fixed bottom-0 left-0 right-0 h-[32px] z-[1] flex justify-center'>
-            <Surface role='hints' limit={1} />
-          </div>
-        )}
-
         {/* Global popovers. */}
         <Popover.Portal>
           <Popover.Content
